Set up task to be scheduled at midnight
