+++ server/liveblog/bandwidth/__init__.py
@@ -1,15 +1,17 @@
 import superdesk
 from liveblog.bandwidth.bandwidth import (
-    Bandwidth,
     bandwidth_key,
     BandwidthResource,
     BandwidthService,
 )
+from .tasks import fetch_bandwidth_usage
+
+__all__ = [
+    "fetch_bandwidth_usage",
+]
 
 
 def init_app(app):
     endpoint_name = bandwidth_key
     service = BandwidthService(endpoint_name, backend=superdesk.get_backend())
     BandwidthResource(endpoint_name, app=app, service=service)
-
-    app.bandwidth = Bandwidth(app, service)
